Removal of the 'default group' concept; removal of MC delete subscriber on remove from group
1. remove the MCAPI delete subscriber call from the remove-from-group
hook. This was causing data loss at Mailchimp - e.g. meaning Mailchimp
could not report how many people unsubscribed as the result of a
campaign, and possibly other bad things.

2. remove the default group concept because this would need to be on a
per-list basis to be useful.

3. require a grouping to be selected when mapping a group to a list.
This may require a "general" grouping or such. This is needed because if
we allow a group to link to a list without a grouping then add a
grouping at the MC end, then it's unclear whether the contact in Civi
should now be linked to the non-grouping group as well as the specified
grouping group.